+++ day3/src/test/java/taink/practice/BatteryBankTest.java
@@ -17,25 +17,25 @@ class BatteryBankTest {
     @Test
     void ascendingBattery() {
         BatteryBank batteryBank = new BatteryBank("234234234234278", 2);
-        assertEquals(78, batteryBank.getMaxJoltage());
+        assertEquals(78L, batteryBank.getMaxJoltage());
     }
 
     @Test
     void descendingBattery() {
         BatteryBank batteryBank = new BatteryBank("987654321111111", 2);
-        assertEquals(98, batteryBank.getMaxJoltage());
+        assertEquals(98L, batteryBank.getMaxJoltage());
     }
 
     @Test
     void spacedBattery() {
         BatteryBank batteryBank = new BatteryBank("811111111111119", 2);
-        assertEquals(89, batteryBank.getMaxJoltage());
+        assertEquals(89L, batteryBank.getMaxJoltage());
     }
 
     @Test
     void randomBattery() {
         BatteryBank batteryBank = new BatteryBank("818181911112111", 2);
-        assertEquals(92, batteryBank.getMaxJoltage());
+        assertEquals(92L, batteryBank.getMaxJoltage());
     }
 
     @Test
@@ -43,8 +43,8 @@ class BatteryBankTest {
         Path path = Paths.get("src/test/resources/small_input.txt");
         List<String> batteries = Files.readAllLines(path);
         List<BatteryBank> batteryBanks2 = batteries.stream().map(s -> new BatteryBank(s, 2)).toList();
-        List<Integer> maxJoltages2 = batteryBanks2.stream().map(BatteryBank::getMaxJoltage).toList();
-        assertArrayEquals(new Integer[] {98, 89, 78, 92}, maxJoltages2.toArray());
-        assertEquals(357, maxJoltages2.stream().reduce(0, Integer::sum));
+        List<Long> maxJoltages2 = batteryBanks2.stream().map(BatteryBank::getMaxJoltage).toList();
+        assertArrayEquals(new Long[] {98L, 89L, 78L, 92L}, maxJoltages2.toArray());
+        assertEquals(357, maxJoltages2.stream().reduce(0L, Long::sum));
     }
 }
